Split up compilation
